requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2.16
djangorestframework==3.14.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.26.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.14
waitress==2.1.2
yarl==1.8.2
